requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.13.3
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.0
asyncio-throttle==1.0.2
python-dotenv==1.0.0
requests==2.31.0 